# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FPS Brandpreventie - Platform voor brandpreventieve gebouwvoorzieningen
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /helpdesk:
    get:
      operationId: listHelpdeskTickets
      tags: [systeem]
      summary: Helpdesktickets (beheerder)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HelpdeskTicket"
    post:
      operationId: createHelpdeskTicket
      tags: [systeem]
      summary: Helpdeskticket indienen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HelpdeskTicketInput"
      responses:
        "201":
          description: Aangemaakt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelpdeskTicket"
  /helpdesk/{id}:
    patch:
      operationId: updateHelpdeskTicket
      tags: [systeem]
      summary: Helpdeskticket-status bijwerken (beheerder)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HelpdeskTicketUpdate"
      responses:
        "200":
          description: Bijgewerkt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelpdeskTicket"
        "404":
          description: Niet gevonden

  # ── FEEDBACK ──────────────────────────────────────────────────────────────
  /feedback:
    get:
      operationId: listFeedback
      tags: [systeem]
      summary: Feedbackoverzicht (beheerder)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Feedback"
    post:
      operationId: createFeedback
      tags: [systeem]
      summary: Feedback indienen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackInput"
      responses:
        "201":
          description: Aangemaakt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"

  # ── HEATMAPS (muisgebeurtenissen) ─────────────────────────────────────────
  /muis-gebeurtenissen:
    get:
      operationId: listMuisGebeurtenissen
      tags: [systeem]
      summary: Muisgebeurtenissen per pagina (beheerder)
      parameters:
        - in: query
          name: pagina
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MuisGebeurtenis"
    post:
      operationId: createMuisGebeurtenissen
      tags: [systeem]
      summary: Batch muisgebeurtenissen vastleggen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MuisGebeurtenisBatch"
      responses:
        "204":
          description: Vastgelegd
...
components:
  schemas:
...
    HelpdeskTicket:
      type: object
      required: [id, onderwerp, bericht, status, aangemaakt_op]
      properties:
        id:
          type: integer
        gebruiker_id:
          type: ["integer", "null"]
        naam:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        onderwerp:
          type: string
        bericht:
          type: string
        status:
          type: string
        aangemaakt_op:
          type: string
    HelpdeskTicketInput:
      type: object
      required: [onderwerp, bericht]
      properties:
        onderwerp:
          type: string
        bericht:
          type: string
    HelpdeskTicketUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
    Feedback:
      type: object
      required: [id, type, bericht, aangemaakt_op]
      properties:
        id:
          type: integer
        gebruiker_id:
          type: ["integer", "null"]
        naam:
          type: ["string", "null"]
        type:
          type: string
        waardering:
          type: ["integer", "null"]
        bericht:
          type: string
        pagina:
          type: ["string", "null"]
        aangemaakt_op:
          type: string
    FeedbackInput:
      type: object
      required: [bericht]
      properties:
        type:
          type: string
        waardering:
          type: integer
        bericht:
          type: string
        pagina:
          type: string
    MuisGebeurtenis:
      type: object
      required: [pagina, type, x, y]
      properties:
        pagina:
          type: string
        type:
          type: string
        x:
          type: number
        y:
          type: number
    MuisGebeurtenisBatch:
      type: object
      required: [gebeurtenissen]
      properties:
        gebeurtenissen:
          type: array
          items:
            $ref: "#/components/schemas/MuisGebeurtenis"
...